backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy>=2.0.36
psycopg2-binary==2.9.9
pgvector==0.3.2
firebase-admin==6.5.0
google-generativeai==0.7.2
stripe==10.5.0
python-dotenv==1.0.1
pytest==8.2.2
pytest-mock==3.14.0
gunicorn==22.0.0
